requirements.txt
ccxt==4.3.89
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b0
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
